Allow drawing 3D models of AruDice
commit-id:caf1c1a1
