+++ packages/marker-detection/src/marker_detection/__init__.py
@@ -1,1 +1,3 @@
+from . import aruco, chessboard
 
+__all__ = ['aruco', 'chessboard']
